Update file names and refactor tool tip wording
