+++ src/components/App/App.tsx
@@ -2,7 +2,7 @@ import React, { useState }from 'react'
 import { Route, Routes } from 'react-router-dom'
 import Nav from '../Header/Header'
 import Home from '../Home/Home'
-import View from '../View/View'
+import View from '../Notice/Notice'
 import Form from '../Form/Form'
 import Preview from '../Preview/Preview'
 import { preview } from '../../preview'
